fixed bug with axios and created cookies

+++ server/src/routes/memories.ts
@@ -4,8 +4,20 @@ import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
 
-    app.get('/memories', async () => {
+    /* We use the hook in case we want everything to pass through the verification */
+    /* We could just put the jwtVerify on the route to be verified */
+    app.addHook('preHandler', async (req) => {
+        await req.jwtVerify()
+    })
+
+    app.get('/memories', async (req) => {
+
+
+
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: req.user.sub
+            },
             orderBy: {
                 createdAt: 'asc'
             }
@@ -23,7 +35,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         })
     })
 
-    app.get('/memories/:id', async (req) => {
+    app.get('/memories/:id', async (req, res) => {
 
         //default way of getting params
         //const {id} = req.params
@@ -35,11 +47,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { id } = paramsSchema.parse(req.params)
 
-        const memory = prisma.memory.findUniqueOrThrow({
+        const memory = await prisma.memory.findUniqueOrThrow({
             where: {
                 id
             }
         })
+
+        if (!memory.isPublic && memory.userId !== req.user.sub) {
+            return res.status(401).send()
+        }
 
         return memory
 
@@ -61,14 +77,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: '7fdd5c7d-a2ac-449d-822c-8249a51fa852',
+                userId: req.user.sub,
             }
         })
 
         return memory
     })
 
-    app.put('/memories/:id', async (req) => {
+    app.put('/memories/:id', async (req, res) => {
         //zod verification
         const paramsSchema = z.object({
             id: z.string().uuid()
@@ -84,7 +100,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { content, coverUrl, isPublic } = bodySchema.parse(req.body)
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id
+            }
+        })
+
+        if (!memory.isPublic && memory.userId !== req.user.sub) {
+            return res.status(401).send()
+        }
+
+        memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -98,7 +124,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory
     })
 
-    app.delete('/memories/:id', async (req) => {
+    app.delete('/memories/:id', async (req, res) => {
         //zod verification
         const paramsSchema = z.object({
             id: z.string().uuid()
@@ -106,14 +132,21 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { id } = paramsSchema.parse(req.params)
 
-        const memory = await prisma.memory.delete({
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id
+            }
+        })
+
+        if (!memory.isPublic && memory.userId !== req.user.sub) {
+            return res.status(401).send()
+        }
+
+        await prisma.memory.delete({
             where: {
                 id
             }
         })
     })
-
-
-
 
 }
